requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# HTTP Requests
requests==2.31.0
urllib3==2.1.0

# Web Scraping (backup)
beautifulsoup4==4.12.2
lxml==5.1.0
selenium==4.16.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Environment Variables
python-dotenv==1.0.0

# Configuration
PyYAML==6.0.1

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2

# AI/LLM Integration
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2

# API Rate Limiting
ratelimit==2.2.1

# Logging
colorlog==6.8.0

# Job Queue (optional for background tasks)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
